+++ utils/const.py
@@ -1,28 +1,51 @@
-from enum import Enum, StrEnum, unique
+"""Constantes, enums y versión de kps."""
 
-Version = "1.3.0" # Mayor.minor.patch
+from enum import StrEnum, unique
+
+Version = "1.3.1"  # pylint: disable=invalid-name  # Mayor.minor.patch; usado por App_version()
+
+# Tiempos por defecto (segundos)
+DEFAULT_AWAY_TIME = 2
+DEFAULT_POLL_INTERVAL = 5
+
+# Rutas y comandos de movimiento del ratón
+MOVE_SCRIPT = "utils/move.py"
+WINDOWS_MOVE_CMD = "cmd /c utils/move.bat"
+
+VENV_DIR_NAME = ".venv"
+
 
 class Display(StrEnum):
+    """Tipo de display GDK detectado."""
+
     def __str__(self):
         return str(self.value)
-    X11 = 'X11Display'
-    WAYLAND = 'GdkWaylandDisplay'
-    WIN32 = 'GdkWin32Display'
-    QUARTZ = 'GdkQuartzDisplay'
+
+    X11 = "X11Display"
+    WAYLAND = "GdkWaylandDisplay"
+    WIN32 = "GdkWin32Display"
+    QUARTZ = "GdkQuartzDisplay"
 
 
 @unique
 class IdleState(StrEnum):
+    """Estado de inactividad del usuario."""
+
     def __str__(self):
         return str(self.value)
-    UNKNOWN = 'OS probably not supported'
-    XA = 'extended away'
-    AWAY = 'away'
-    AWAKE = 'awake'
+
+    UNKNOWN = "OS probably not supported"
+    XA = "extended away"
+    AWAY = "away"
+    AWAKE = "awake"
+
 
 @unique
 class OsType(StrEnum):
+    """Identificador de familia OS (os.name)."""
+
     def __str__(self):
         return str(self.value)
-    UNIX = 'posix'
-    WINDOWS = 'nt'
+
+    UNIX = "posix"
+    WINDOWS = "nt"